requirements.txt
playwright==1.55.0
pytest==8.4.2
pytest-playwright==0.5.2
python-dotenv==1.0.1
pytest-xdist==3.7.0 
pytest-html==3.2.0
